refac: move balance calc back to domain model
The reason why the calculation was placed directly under a query handler is because the balance was not expected be a part of domain validation, say, during the processing of a command, and the retrieve_account query won't necessarily continue to load an account entity moving forward. That said, at the stage the project is, having the balance calculation provided by the Account class makes the code more readable, and it can change later as needed.